requirements.txt
# Occupancy Detection Service — Dependencies
# ต้องใช้ Python 3.10 ขึ้นไป

# ── Core ──
ultralytics>=8.0
opencv-python-headless>=4.8
numpy>=1.24

# ── API Server ──
fastapi>=0.100
uvicorn[standard]>=0.22
python-multipart>=0.0.6        # สำหรับ file upload

# ── Config ──
pyyaml>=6.0

# ── MQTT (optional — ถ้าไม่ใช้ MQTT ลบบรรทัดนี้ได้) ──
# pin <2.0 เพราะ paho-mqtt v2 เปลี่ยน API (Client init ต้องระบุ callback_api_version)
paho-mqtt>=1.6,<2.0
